requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-bcrypt>=1.0.1
flask-jwt-extended>=4.5.3
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
openai>=1.3.0
boto3>=1.34.0
twilio>=8.12.0
cryptography>=42.0.0
flask-limiter>=3.5.0
flask-talisman>=1.1.0
gunicorn>=21.2.0